// burak-react/src/lib/enums/common.enum.ts
export enum ProductCollection {
  DISH = "DISH",
  SALAD = "SALAD",
  DESSERT = "DESSERT",
  DRINK = "DRINK",
  OTHER = "OTHER",
}

export enum ProductSize {
  SMALL = "SMALL",
  NORMAL = "NORMAL",
  LARGE = "LARGE",
  SET = "SET",
}

export enum ProductStatus {
  PAUSE = "PAUSE",
  PROCESS = "PROCESS",
  DELETE = "DELETE",
}

export enum MemberType {
  USER = "USER",
  RESTAURANT = "RESTAURANT",
}

export enum MemberStatus {
  ACTIVE = "ACTIVE",
  BLOCK = "BLOCK",
  DELETE = "DELETE",
}

export enum OrderStatus {
  PAUSE = "PAUSE",
  PROCESS = "PROCESS",
  FINISH = "FINISH",
  DELETE = "DELETE",
}